Fix: Arg Min/Max reduction patches post public API refactor (#30)
* fix: argmin/max 1 pass fixes after public API refactor

* fix jit.jl test causing failure

+++ benchmarks/main-benchmarks/baseline/linreg/Param.h
@@ -12,7 +12,7 @@
 
 typedef int32_t T;
 #ifndef RED_T
-typedef int64_t RED_T;
+typedef int32_t RED_T;
 #endif
 
 const uint32_t check_correctness = 0;
